Add image handling to blog posts and update styles
This commit incorporates featured images and captions into blog posts. It also enhances the styles for the images displayed on the home and blog post pages. The change allows featured images to be included in each individual blog post markdown file and subsequently displayed on the blog post page and home page, which will enrich the visual appeal of the blog. Lastly, corresponding styles in the main.scss file have been updated to refine the appearance of featured images, such as hover transformations and border-radius properties.