Update path and prefix for Gatsby deployment
This commit modifies gatsby-config.js to add pathPrefix and assetPrefix for deployment purposes. There are also changes to the package.json file, specifically to the "build" and "deploy" scripts to include the "--prefix-paths" option. This ensures that assets and pages are correctly served when the site is hosted on GitHub pages.